BE_MojiChat/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-socketio